feat: 목표달성 반영하여 프로그레스뷰 완성

+++ iOS_Wegg/iOS_Wegg/Screen/Models/NetworkModel/Todo/TodoResponse.swift
@@ -28,7 +28,7 @@ struct TodoAchievementResponse: Decodable {
     let isSuccess: Bool
     let code: String
     let message: String
-    let result: Int
+    let result: Double
 }
 
 /// Todo리스트 조회 response
